mines 4

+++ tasks/mines/minesweeper.h
@@ -49,7 +49,7 @@ public:
     std::set<std::pair<size_t, size_t>> GenerateMines(size_t width, size_t height, size_t mines_count) const;
 
 private:
-    int closed_;
+    size_t closed_;
     Information field_;
     GameStatus status_ = GameStatus::NOT_STARTED;
     time_t time_ = std::time(0);
